+++ types/factories/IUniswapV3PoolForTest__factory.ts
@@ -27,9 +27,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "value",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "Approval",
@@ -64,15 +64,15 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "Burn",
@@ -169,27 +169,27 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "paid0",
-        type: "uint",
+        type: "uint256",
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "paid1",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "Flash",
@@ -268,15 +268,15 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "Mint",
@@ -379,9 +379,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "value",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "Transfer",
@@ -403,9 +403,9 @@ const _abi = [
     name: "allowance",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -419,9 +419,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "approve",
@@ -446,9 +446,9 @@ const _abi = [
     name: "balanceOf",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -475,14 +475,14 @@ const _abi = [
     name: "burn",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "nonpayable",
@@ -597,9 +597,9 @@ const _abi = [
     name: "feeGrowthGlobal0X128",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -610,9 +610,9 @@ const _abi = [
     name: "feeGrowthGlobal1X128",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -626,14 +626,14 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
       {
         internalType: "bytes",
@@ -729,14 +729,14 @@ const _abi = [
     name: "mint",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount0",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount1",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "nonpayable",
@@ -745,9 +745,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "index",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "observations",
@@ -816,14 +816,14 @@ const _abi = [
         type: "uint128",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthInside0LastX128",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthInside1LastX128",
-        type: "uint",
+        type: "uint256",
       },
       {
         internalType: "uint128",
@@ -1007,9 +1007,9 @@ const _abi = [
     name: "tickBitmap",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -1049,14 +1049,14 @@ const _abi = [
         type: "int128",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthOutside0X128",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthOutside1X128",
-        type: "uint",
+        type: "uint256",
       },
       {
         internalType: "int56",
@@ -1116,9 +1116,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "transfer",
@@ -1145,9 +1145,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "transferFrom",
