+++ types/factories/IUniswapV3PoolState__factory.ts
@@ -15,9 +15,9 @@ const _abi = [
     name: "feeGrowthGlobal0X128",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -28,9 +28,9 @@ const _abi = [
     name: "feeGrowthGlobal1X128",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -52,9 +52,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "index",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "observations",
@@ -99,14 +99,14 @@ const _abi = [
         type: "uint128",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthInside0LastX128",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthInside1LastX128",
-        type: "uint",
+        type: "uint256",
       },
       {
         internalType: "uint128",
@@ -194,9 +194,9 @@ const _abi = [
     name: "tickBitmap",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -223,14 +223,14 @@ const _abi = [
         type: "int128",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthOutside0X128",
-        type: "uint",
+        type: "uint256",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "feeGrowthOutside1X128",
-        type: "uint",
+        type: "uint256",
       },
       {
         internalType: "int56",
