+++ types/factories/Referral__factory.ts
@@ -61,9 +61,9 @@ const _abi = [
       },
       {
         indexed: false,
-        internalType: "uint",
+        internalType: "uint256",
         name: "commission",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "ReferralCommissionRecorded",
@@ -99,9 +99,9 @@ const _abi = [
     name: "countReferrals",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -115,9 +115,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "_debit",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "debitOutstandingCommission",
@@ -136,9 +136,9 @@ const _abi = [
     name: "getOutstandingCommission",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "amount",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -174,9 +174,9 @@ const _abi = [
     name: "outstandingCommissions",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
@@ -234,9 +234,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "_commission",
-        type: "uint",
+        type: "uint256",
       },
     ],
     name: "recordReferralCommission",
@@ -294,9 +294,9 @@ const _abi = [
     name: "totalReferralCommissions",
     outputs: [
       {
-        internalType: "uint",
+        internalType: "uint256",
         name: "",
-        type: "uint",
+        type: "uint256",
       },
     ],
     stateMutability: "view",
